Add ability to use exporter without id property
The id property of the exporter is used for checking the default values
specified in `prometheus_target_exporter_defaults` variable.
However when both the `host` and the `path` fields are specified
manually in the `prometheus_target_exporter` list this id is never
looked up.
The way the templating was handled however, still resulted in an error
when leaving out the id field.
The templating code was now rewritten to allow for the following
exporter configuraion to run:

prometheus_target_exporter:
- { host: exporter_without_id, path: /opt/simple_target4.yml }

Signed-off-by: Colin Wilk <colin.wilk@tum.de>