Add run_once option for handlers
This adds the run_once option for the two handler types (shell and
command). The variables `prometheus_target_handler_command_run_once`
and `prometheus_target_handler_shell_run_once` are introduced to
control this behavior.

For this change a switch from the Ansible Handlers to regular tasks was
necessary to  enforce run_once behavior with delegated hosts. If you
don't rely on handler specific features such as deferring, nothing
changes when using the role.

Signed-off-by: Colin Wilk <colin.wilk@tum.de>